requirements.txt
numpy
scipy
jieba
sklearn
nltk
flask
flask-mongoengine
